fixed some small minor issues
