+++ mobile/app/src/main/java/com/example/timed_mobile/TimeOutActivity.kt
@@ -84,44 +84,3 @@ class TimeOutActivity : AppCompatActivity() {
         }
     }
 }
-
-
-// ✅ ExcuseLetterActivity.kt – Full Firebase integration example
-package com.example.timed_mobile
-
-import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.database.FirebaseDatabase
-
-class ExcuseLetterActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.excuse_letter_page)
-
-        val reasonInput = findViewById<EditText>(R.id.reason_input)
-        val submitButton = findViewById<Button>(R.id.btn_submit_excuse)
-
-        submitButton.setOnClickListener {
-            val reason = reasonInput.text.toString().trim()
-            val userId = FirebaseAuth.getInstance().currentUser?.uid
-            val dbRef = FirebaseDatabase.getInstance().getReference("excuseLetters")
-
-            val excuseData = mapOf(
-                "userId" to userId,
-                "reason" to reason,
-                "timestamp" to System.currentTimeMillis()
-            )
-
-            dbRef.push().setValue(excuseData).addOnSuccessListener {
-                Toast.makeText(this, "Excuse submitted", Toast.LENGTH_SHORT).show()
-                reasonInput.text.clear()
-            }.addOnFailureListener {
-                Toast.makeText(this, "Submission failed: ${it.message}", Toast.LENGTH_SHORT).show()
-            }
-        }
-    }
-}
