+++ mobile/app/src/main/res/drawable/edittext_background.xml
@@ -1,0 +1,6 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#F3F3F3" />
+    <stroke android:width="1dp" android:color="#CCCCCC"/>
+    <corners android:radius="12dp"/>
+</shape>
