+++ mobile/app/src/main/res/layout/excuse_letter_page.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:padding="24dp"
+    android:background="#FFFFFF">
+
+    <!-- Title -->
+    <TextView
+        android:id="@+id/excuse_title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Submit Excuse Letter"
+        android:textSize="24sp"
+        android:textStyle="bold"
+        android:textColor="#8B0000"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginTop="32dp" />
+
+    <!-- Reason Input Field -->
+    <EditText
+        android:id="@+id/reason_input"
+        android:layout_width="0dp"
+        android:layout_height="200dp"
+        android:hint="State your reason here..."
+        android:gravity="top|start"
+        android:inputType="textMultiLine"
+        android:background="@drawable/edittext_background"
+        android:padding="16dp"
+        android:textSize="16sp"
+        app:layout_constraintTop_toBottomOf="@id/excuse_title"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginTop="24dp" />
+
+    <!-- Submit Button -->
+    <Button
+        android:id="@+id/btn_submit_excuse"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:text="Submit"
+        android:backgroundTint="#FFC107"
+        android:textColor="#8B0000"
+        android:textStyle="bold"
+        android:elevation="6dp"
+        android:layout_marginTop="24dp"
+        app:cornerRadius="12dp"
+        app:layout_constraintTop_toBottomOf="@id/reason_input"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
